requirements.txt
# Slack API
slack-sdk==3.26.1

# Google Sheets API
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# Environment variables
python-dotenv==1.0.0

# Data processing
pandas==2.1.4

# Logging
colorlog==6.8.0
